pyproject.toml
[project]
name = "neurommind-mapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Joe McMullin", email = "joemcmullin@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.62.0",
    "beautifulsoup4>=4.13.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.48.0",
]

[project.scripts]
neurommind-mapper = "neurommind_mapper:main"

[build-system]
requires = ["uv_build>=0.8.8,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
]
